separated gc_allocator and allocator
