+++ src/raw_allocator/error.rs
@@ -1,0 +1,5 @@
+#[derive(Debug)]
+pub enum AllocError {
+    OOM,
+    AllocOverflow,
+}
