+++ src/raw_allocator/mod.rs
@@ -6,8 +6,10 @@ mod block_store;
 mod bump_block;
 mod constants;
 mod size_class;
+mod error;
 
 #[cfg(test)]
 mod tests;
 
-pub use allocator::GcAllocator;
+pub use allocator::Allocator;
+pub use error::AllocError;
